Refactor sensor and communication data types, remove unused files
Changed motion and battery data types from float to int for Firebase transmission. Simplified LED control logic and removed state tracking for PWM updates. Deleted unused wifi_serial and debug header files. Updated main setup and loop to remove WiFi connection wait and enable UART acknowledgment and communication failsafe logic.

+++ ESP32-S3-Main/src/communication/firebase_manager.h
@@ -6,10 +6,10 @@
 
 struct FirebaseTxData {
     // Sensor readings - SEND ONLY
-    float acceleration;           // Send acceleration
+    int acceleration;           // Send acceleration
     int angular;               // Send angular velocity
     int battery;               // Send battery level
-    float voltage;             // Send voltage
+    int voltage;               // Send voltage
     int temp;                  // Send temperature
     int humidity;              // Send humidity
     int lightlevel;            // Send light amount from path LDRs
